Commit lần cuối sau khi báo cáo xong

+++ backend/controllers/card.controller.js
@@ -1,17 +1,14 @@
-const CardService = require('../services/card.service');
 const http = require('http-status-codes');
 const objectName = require('../utils/props-and-objects.util').OBJECTS.CARD;
 const actionCreate = require('../utils/props-and-objects.util').ACTIONS.CREATE;
 const actionUpdate = require('../utils/props-and-objects.util').ACTIONS.UPDATE;
 const actionDelete = require('../utils/props-and-objects.util').ACTIONS.DELETE;
-const frontTextProp = require('../utils/props-and-objects.util').PROPS.FRONT_TEXT;
-const backTextProp = require('../utils/props-and-objects.util').PROPS.BACK_TEXT;
-const DECK = require('../utils/props-and-objects.util').OBJECTS.DECK;
 const requestMessageUtil = require('../utils/requestMessage.util');
 const cardService = require('../services/card.service');
+const accountCardDetailService = require('../services/account-card-detail.service');
 
 const CardController = {
-  getAllCards: async (req, res) => {
+  getAllCards: async (_, res) => {
     try {
       const cards = await cardService.getAllCards();
       res.status(http.StatusCodes.OK).json(cards);
@@ -23,10 +20,12 @@ const CardController = {
   createCard: async (req, res) => {
     try {
       const cardData = req.body;
-      console.log(cardData);
-      const front_image = req.file;
-      const card = await cardService.createCard(cardData, front_image);
-
+      // console.log(cardData);
+      let { front_image_file, front_audio_file } = req.files;
+      front_image_file = front_image_file ? front_image_file[0] : null;
+      front_audio_file = front_audio_file ? front_audio_file[0] : null;
+      // console.log(front_image_file, front_audio_file);
+      await cardService.createCard(cardData, front_image_file, front_audio_file);
       res.status(http.StatusCodes.CREATED).json({
         message: requestMessageUtil.successActionObject(actionCreate, objectName),
       });
@@ -38,17 +37,13 @@ const CardController = {
   updateCard: async (req, res) => {
     try {
       const { id } = req.params;
-      const { front_text, back_text, deck_id, is_public, word_class_id } = req.body;
-      const front_image = req.file;
-      const updatedCard = await cardService.updateCard(
-        id,
-        front_text,
-        front_image,
-        back_text,
-        deck_id,
-        is_public,
-        word_class_id
-      );
+      const cardData = req.body;
+      let { front_image_file, front_audio_file } = req.files;
+      console.log(front_image_file, front_audio_file);
+      front_image_file = front_image_file ? front_image_file[0] : null;
+      front_audio_file = front_audio_file ? front_audio_file[0] : null;
+      // console.log(front_image_file, front_audio_file);
+      const updatedCard = await cardService.updateCard(id, cardData, front_image_file, front_audio_file);
       if (!updatedCard) {
         return res.status(http.StatusCodes.NOT_FOUND).json({
           error: requestMessageUtil.notFoundObject(objectName),
@@ -66,6 +61,12 @@ const CardController = {
   deleteCard: async (req, res) => {
     try {
       const { id } = req.params;
+      const accountCardDetail = accountCardDetailService.getAccountCardDetails();
+      if (accountCardDetail) {
+        return res.status(http.StatusCodes.FORBIDDEN).json({
+          error: 'Không thẻ xóa thẻ đã có người học',
+        });
+      }
       const deletedCard = await cardService.deleteCard(id);
       if (!deletedCard) {
         return res.status(http.StatusCodes.NOT_FOUND).json({
